+++ crates/repos/src/lib.rs
@@ -2,6 +2,7 @@ pub mod entity;
 pub mod error;
 pub mod postgres;
 pub mod repo;
+pub mod urn;
 
 use crate::{
     postgres::{PostgresEpisodeRepo, PostgresScriptRepo},
