+++ crates/repos/src/postgres.rs
@@ -1,10 +1,10 @@
 use crate::{
-    entity::{Comment, Episode, Podcast, Script, Task, TaskStatus},
-    error::Error,
-    repo::{
-        CommentId, CommentRepo, EpisodeId, EpisodeRepo, PodcastId, PodcastRepo, ScriptId,
-        ScriptRepo, TaskId, TaskRepo,
+    entity::{
+        Comment, CommentId, Episode, EpisodeId, Podcast, PodcastId, Script, ScriptId, Task, TaskId,
+        TaskStatus,
     },
+    error::Error,
+    repo::{CommentRepo, EpisodeRepo, PodcastRepo, ScriptRepo, TaskRepo},
 };
 use async_trait::async_trait;
 use chrono::{Local, Utc};
