+++ crates/script_runtime/src/imports/urn.rs
@@ -1,11 +1,10 @@
-use crate::Urn;
+use crate::resolve::resolve_urn;
 use anyhow::Result;
 use json_e::{
     value::{AsyncCallable, Value},
     Context,
 };
-use repos::{comment_repo, episode_repo, error::Error, podcast_repo, repo::PodcastId, script_repo};
-use uuid::Uuid;
+use repos::urn::Urn;
 
 #[derive(Clone)]
 pub(crate) struct Eval;
@@ -38,35 +37,7 @@ impl AsyncCallable for UrnGet {
             [Value::String(urn)] => urn.parse::<Urn>(),
             _ => return Err(anyhow::anyhow!("invalid args".to_string())),
         }?;
-
-        let value = match urn {
-            Urn::Podcast(id) => {
-                let podcast_repo = podcast_repo();
-                let podcast = podcast_repo.find_by_id(&id).await?;
-                serde_json::to_value(podcast)
-            }
-            Urn::Episode(id) => {
-                let episode_repo = episode_repo();
-                let (episode, comments) = episode_repo.find_by_id(&id).await?;
-                let mut episode = serde_json::to_value(episode)?;
-                episode
-                    .as_object_mut()
-                    .unwrap()
-                    .insert("comments".to_string(), serde_json::to_value(comments)?);
-                Ok(episode)
-            }
-            Urn::Comment(id) => {
-                let comment_repo = comment_repo();
-                let res = comment_repo.find_by_id(&id).await?;
-                serde_json::to_value(res)
-            }
-            Urn::Script(id) => {
-                let script_repo = script_repo();
-                let res = script_repo.find_by_id(&id).await?;
-                serde_json::to_value(res)
-            }
-            Urn::Other(resource, id) => return Err(Error::NotFound(resource, id).into()),
-        }?;
+        let value = resolve_urn(urn).await?;
         Ok(value.into())
     }
 }
