+++ crates/script_runtime/src/resolve.rs
@@ -1,0 +1,34 @@
+use anyhow::Result;
+use repos::{comment_repo, episode_repo, error::Error, podcast_repo, script_repo, urn::Urn};
+
+pub(crate) async fn resolve_urn(urn: Urn) -> Result<serde_json::Value> {
+    let value = match urn {
+        Urn::Podcast(id) => {
+            let podcast_repo = podcast_repo();
+            let podcast = podcast_repo.find_by_id(&id).await?;
+            serde_json::to_value(podcast)?
+        }
+        Urn::Episode(id) => {
+            let episode_repo = episode_repo();
+            let (episode, comments) = episode_repo.find_by_id(&id).await?;
+            let mut episode = serde_json::to_value(episode)?;
+            episode
+                .as_object_mut()
+                .unwrap()
+                .insert("comments".to_string(), serde_json::to_value(comments)?);
+            episode
+        }
+        Urn::Comment(id) => {
+            let comment_repo = comment_repo();
+            let res = comment_repo.find_by_id(&id).await?;
+            serde_json::to_value(res)?
+        }
+        Urn::Script(id) => {
+            let script_repo = script_repo();
+            let res = script_repo.find_by_id(&id).await?;
+            serde_json::to_value(res)?
+        }
+        Urn::Other(resource, id) => return Err(Error::NotFound(resource, id).into()),
+    };
+    Ok(value)
+}
